refactor: move layouts to config property
