+++ examples/web/src/pages/ClassicApiReferencePage.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import {
-  ApiReferenceClassic,
+  ApiReference,
   type ReferenceConfiguration,
 } from '@scalar/api-reference'
 import { reactive } from 'vue'
@@ -9,15 +9,15 @@ const configuration = reactive<ReferenceConfiguration>({
   theme: 'default',
   proxy: 'http://localhost:5051',
   isEditable: false,
+  layout: 'classic',
   spec: {
     url: 'https://petstore.swagger.io/v2/swagger.json',
   },
-  showSidebar: false,
   tabs: {
     initialContent: 'Swagger Editor',
   },
 })
 </script>
 <template>
-  <ApiReferenceClassic :configuration="configuration" />
+  <ApiReference :configuration="configuration" />
 </template>
